Inventory and equipment update
The bulk of this update is around using items in and out of combat from the inventory screen. There is an eq% (equipped?) flag in the items database which I couldn't get to work as I wanted (skill issue) so instead equipped items are stored in GAME in string variables and assured by code logic... I'd like to go back and make the database based equipped logic work.